inventory_management_system/task/requirement.txt
prettytable==3.17.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
wcwidth==0.6.0